Splits the JCommander code into it's own class to allow embedded applications to exclude it
